requirements.txt
APScheduler==3.7.0
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Deprecated==1.2.12
fire==0.3.1
Flask==1.1.2
Flask-APScheduler==1.12.1
Flask-Script==2.0.6
idna==2.10
ioc-fanger==3.2.3
ioc-finder==5.0.2
iocextract==1.13.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kafka-python==2.0.2
markdown-tree-parser==0.1
MarkupSafe==1.1.1
mdutils==1.3.0
pdfminer==20191125
pendulum==1.2.5
Pillow==8.2.0
pyattck==3.0.1
pycountry==20.7.3
pycryptodome==3.10.1
pyfiglet==0.8.post1
pymisp==2.4.141.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-gitlab==2.6.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.2.1
termcolor==1.1.0
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
